+++ CHANGES.rst
@@ -5,10 +5,12 @@ These features will be included in the next release:
 
 Added
 -----
+- Unit test for ``log_level`` option for `darkgraylib.config.load_config`.
 
 Fixed
 -----
 - Keep Pylint below version 3.3.0 until we drop support for Python 3.8.
+- Don't convert ``log_level = "<string>"`` in the configuration file to an integer.
 
 
 2.0.1_ - 2024-08-09